pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-seat_swap_inference"
version = "0.1.0"
description = "Seat Swap Inference environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "google-generativeai>=0.8.6",
    # Core OpenEnv runtime (provides FastAPI server + HTTP client types)
    # install from github
    # "openenv-core[core] @ git+https://github.com/meta-pytorch/OpenEnv.git",
    "openenv-core[core]>=0.2.2",
    # Environment-specific dependencies
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m seat_swap_inference.server.app
server = "seat_swap_inference.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["seat_swap_inference", "seat_swap_inference.server"]
package-dir = { "seat_swap_inference" = ".", "seat_swap_inference.server" = "server" }
